requirements.txt
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
Flask==3.1.0
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
pycryptodome==3.22.0
PyMySQL==1.1.1
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
